v0.1.4
Updated io.py to include key parsing as an option when loading the database.
Overloaded so previous behaviour does not produce an error.

All scripts updated to do key parsing using built in load_db. Note this also
allows for the negating of keys in the search using the ~ operator.

Signed-off-by: Andrew Schaeffer <aj.schaeffer@gmail.com>