v0.2.0
Added Long Key formatting and proper construction and handling of channel and location ids in query_fdsn_stdb.py.
Updated channel ranking to specify HH>BH>LH, and without long format keys, the first channel in the ranking and
request found will be stored. The date in the stdb will also reflect the date of the channel, not the date of the
station operation time (which may include a wider window from other channels).



Signed-off-by: Andrew Schaeffer <aj.schaeffer@gmail.com>; 2020-12-01